Move MoviesService.cs to Jellyfin.Api
